requirements.txt
opencv-python==4.0.0.21
imutils==0.5.2
numpy==1.16.2
torch==1.0.1
torchvision==0.2.2.post3
pillow==5.4.1